- Resolve some Comments from Review - Use arrange-act-assert pattern for unit Tests - split test files for OptionExtensions and ResultExtensions with same pattern.
